# use fast base58 encode implementation (20~60% performance gain)