```
$ go test -bench=. base58/base58bench_test.go 
BenchmarkBase58FastEncode-7        	      30	  45674986 ns/op	   0.11 MB/s
BenchmarkBase58SlowEncode-7        	      20	  63573231 ns/op	   0.08 MB/s
BenchmarkBase58FastEncodeSmall-7   	 1000000	      1712 ns/op	  17.52 MB/s
BenchmarkBase58SlowEncodeSmall-7   	  300000	      5336 ns/op	   5.62 MB/s
BenchmarkBase58Decode-7            	     300	   5276054 ns/op	   1.29 MB/s
```