@jrick 2017-04-19
And while it's useful for the benchmarks, now that we've seen them, there's no reason to keep the slow implementation around.