@stevenroose 2017-04-28
Indeed. A `FastEncode` method is really not necessary. Just replace the existing encode implementation with the fast one.