@davecgh 2017-05-09
Thanks for the PR.  As others have mentioned, I'd really prefer just replacing the `Encode` with the new faster method.  If you would please do that and rebase/squash everything down, I'll get this approved and merged.