@rubensayshi 2017-05-11
@davecgh okay, glad to see it's well received, I'll cleanup the PR ;)