@rubensayshi 2017-05-12
whoops, almost forgot the important stuff!  
the implementation is from https://github.com/mr-tron/go-base58 MIT license 

how to properly incorperate this?