@jrick 2017-05-12
For the license, you can simply copy it into the source file.