@davecgh 2017-05-12
Oh, I didn't realize it had another license.  Everything in btcsuite must be licensed under the ISC license.  Unfortunately, I'm going to have to decline it as is unless you can get permission to relicense it.