@stevenroose 2017-05-12
@davecgh Isn't the ISC license totally equivalent to the MIT license?

> The ISC license is a permissive free software license written by the Internet Systems Consortium (ISC). It is meant to be functionally equivalent to the simplified BSD and the MIT licenses, differing in its removal of language deemed unnecessary following the global adoption of the Berne Convention.