@rubensayshi 2017-06-22
@davecgh @mr-tron here is the original author, is that enough or you need some more explicit?