@davecgh 2017-06-22
Great.  Thanks @mr-tron.

@rubensayshi Thanks for following up.  The author has provided permission.  That's all we needed.