@ribasushi 2020-06-02
For those following along: I opened a new PR based on a newer iteration of pretty much the same code. Enc/Dec of b58 addresses is 4+/7+ **times** faster: https://github.com/btcsuite/btcutil/pull/170#issue-426748886