@ribasushi 2020-06-03
@rubensayshi I am sorry your PR got stuck, not sure why it happened :(
If time permits and you happen to have a more extensive test suite: it would be great if you test the proposed changes a bit more. You know... just to be sure ;)